[BE] feat: typeORM 연동 및 App.ts 전환 #95
